changed get duration result calc

+++ include/delay.h
@@ -1,5 +1,5 @@
 #include "headers.h"
 
 void delay_init(void);
-uint32_t get_duration(void);
+uint16_t get_duration(void);
 void delay(bool div);
